+++ mission-fe/src/assets/icons/img_default.png
