+++ mission-fe/src/assets/images/home/hero-image.png
